+++ spec/rubocop/cop/layout/indent_assignment_spec.rb
@@ -105,4 +105,20 @@ RSpec.describe RuboCop::Cop::Layout::IndentAssignment, :config do
         baz = ''
     RUBY
   end
+
+  it 'registers an offense for incorrectly indented rhs when' \
+     'multiple assignment with line breaks on each line' do
+    expect_offense(<<~RUBY)
+      foo =
+        bar =
+        baz = 42
+        ^^^^^^^^ Indent the first line of the right-hand-side of a multi-line assignment.
+    RUBY
+
+    expect_correction(<<~RUBY)
+      foo =
+        bar =
+          baz = 42
+    RUBY
+  end
 end